updated Manager and StoreTest
Still don't have a method to check the total sales from the music store.
Also, I'm not sure if I'm checking the inventory correctly. I tried to
add another movie and then query the inventory but got back a strange
result. It looks like a memory address and not the name of the actual
movie/movies in the inventory.